# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Intelligent Teacher Recruitment Platform API",
    "version": "1.0.0",
    "description": "Backend REST API for Intelligent Teacher Recruitment Platform featuring AI matching, application parsing, and role-based recruitment workflows."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide your JWT access token (Bearer <token>) to authenticate requests."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List job postings with optional filters",
        "responses": {
          "200": {
            "description": "List of job postings"
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create a new job posting",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 180
                  },
                  "description": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 10000
                  },
                  "requirements": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 10000
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "DRAFT",
                      "PUBLISHED",
                      "CLOSED",
                      "ARCHIVED"
                    ],
                    "default": "DRAFT"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "title",
                  "description",
                  "requirements"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job post created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}/apply": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "Submit job application with CV document upload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Application submitted successfully"
          },
          "400": {
            "description": "Validation error or invalid document"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/applications": {
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "List submitted job applications",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of applications"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/ai-models": {
      "get": {
        "tags": [
          "AI Models"
        ],
        "summary": "List registered AI matching models",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of AI models"
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List all platform users with filtering",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated users list"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}